requirements.txt
prometheus_client~=0.22.0
requests~=2.32.3
PyYAML~=6.0.2
